Separate local and live tracker integrations
