+++ .github/workflows/tracker-integration-tests.yaml
@@ -10,7 +10,33 @@ permissions:
   id-token: write
 
 jobs:
-  tracker-integration-tests:
+  tracker-local-integration-tests:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+
+      - uses: actions/setup-python@v5
+        with:
+          python-version: "3.12"
+
+      - name: Install uv
+        uses: astral-sh/setup-uv@v5
+        with:
+          enable-cache: true
+          cache-dependency-glob: "services/tracker/uv.lock"
+
+      - name: Install tracker dependencies
+        run: |
+          cd services/tracker
+          uv venv --python 3.12
+          uv sync --group dev
+
+      - name: Run local integration tests
+        run: |
+          cd services/tracker
+          make test-integration-local
+
+  tracker-live-integration-tests:
     runs-on: ubuntu-latest
     steps:
       - name: Skip automatic runs on PRs
@@ -47,10 +73,10 @@ jobs:
           role-to-assume: arn:aws:iam::613431292675:role/github-actions-valkyrie-tests
           aws-region: us-east-1
 
-      - name: Run integration tests
+      - name: Run live integration tests
         run: |
           cd services/tracker
-          uv run pytest tests/integration -n auto
+          uv run pytest tests/integration/live -n auto
         env:
           TEST_AWS_S3_BUCKET: ${{ secrets.TEST_AWS_S3_BUCKET }}
           TEST_LOG_GROUP: ${{ secrets.TEST_LOG_GROUP }}
