+++ services/tracker/tests/integration/live/sandbox/test_provider.py
@@ -1,3 +1,5 @@
+"""Live sandbox-provider lifecycle tests."""
+
 import asyncio
 from collections.abc import Awaitable, Callable
