+++ services/tracker/tests/integration/live/storage/test_upload_artifacts_images.py
@@ -1,3 +1,5 @@
+"""Live artifact image upload tests."""
+
 import asyncio
 from uuid import uuid4
