+++ services/tracker/tests/integration/live/storage/test_upload_to_s3.py
@@ -1,3 +1,5 @@
+"""Live S3 archive upload tests."""
+
 import asyncio
 import io
 import tarfile
