--- services/tracker/tests/integration/test_database_integration.py
@@ -1,346 +1,0 @@
-import logging
-import uuid
-from asyncio import Semaphore, gather
-from random import sample
-from typing import Any
-from uuid import UUID
-
-import pytest
-from benchmark_service import SandboxProvider
-from benchmark_service.client import BenchmarkServiceClient
-from benchmark_service.schemas import FinalScoreResponse, RetrieveTaskResponse
-from sqlalchemy.engine import Engine
-from sqlmodel import Session, col, inspect, select
-
-from tests.utils import random_task_id
-from tests.conftest import TEST_ORG_ID
-from tracker.database.models import (
-    Benchmark,
-    BenchmarkStatus,
-    EvaluationResult,
-    FinalEvaluation,
-    Task,
-    TaskStatus,
-)
-from tracker.sandbox import create_sandbox
-
-logger = logging.getLogger(__name__)
-
-
-class TestDatabaseIntegration:
-    async def _create_task(self, database_session: Session, task_id: str, benchmark_id: UUID) -> Task:
-        task_row = Task(org_id=TEST_ORG_ID, task_id=task_id, benchmark=benchmark_id)
-        database_session.add(task_row)
-        database_session.flush()
-
-        return task_row
-
-    async def _create_evaluation_result(
-        self, database_session: Session, task_row: Task, evaluation_result: dict[str, Any]
-    ) -> EvaluationResult:
-        evaluation_result_row = EvaluationResult(
-            org_id=TEST_ORG_ID, task=task_row.id, instance_id=str(uuid.uuid4()), result=evaluation_result
-        )
-        database_session.add(evaluation_result_row)
-
-        task_row.status = TaskStatus.FINISHED
-        database_session.add(task_row)
-        database_session.flush()
-
-        return evaluation_result_row
-
-    async def _create_final_evaluation(
-        self, database_session: Session, benchmark_row: Benchmark, final_score_response: FinalScoreResponse
-    ) -> FinalEvaluation:
-        final_evaluation_row = FinalEvaluation(
-            org_id=TEST_ORG_ID,
-            benchmark=benchmark_row.id,
-            final_score=final_score_response.final_score,
-            properties=final_score_response.metadata,
-        )
-        database_session.add(final_evaluation_row)
-        database_session.flush()
-
-        return final_evaluation_row
-
-    async def test_create_tables(self, postgres_engine: Engine):
-        """Verify the integration database schema is created with tracker tables.
-
-        Test cases:
-        - The database contains at least one created table.
-        - Benchmark, task, evaluation result, and final evaluation tables are present.
-        """
-        inspector = inspect(postgres_engine)
-        tables = inspector.get_table_names()
-
-        assert len(tables) > 0, "Tables were not created in the database as expected"
-
-        # Verify expected tables exist
-        expected_tables = {"benchmark", "task", "evaluationresult", "finalevaluation"}
-        assert expected_tables.issubset(set(tables)), f"Missing tables. Found: {tables}"
-
-    async def test_database_integrity(self, database_session: Session, example_benchmark_object: Benchmark):
-        """Verify status changes maintain finished_at timestamps on persisted rows.
-
-        Test cases:
-        - Benchmark and task rows start without finished_at while still active.
-        - FINISHED and ERROR statuses set finished_at for both benchmark and task rows.
-        """
-
-        # Test the benchmark table
-        benchmark_row = example_benchmark_object
-        database_session.add(benchmark_row)
-
-        # When created its in pending status
-        assert benchmark_row.status == BenchmarkStatus.IN_PROGRESS
-        assert benchmark_row.finished_at is None
-
-        # When the status is updated to finished, the finished_at timestamp should be set
-        benchmark_row.status = BenchmarkStatus.FINISHED
-        database_session.add(benchmark_row)
-        database_session.flush()
-        assert benchmark_row.finished_at, "Should be auto generated when the status is updated to finished"
-
-        # Test the task table
-        task_row = Task(org_id=TEST_ORG_ID, task_id=random_task_id(), benchmark=benchmark_row.id)
-        database_session.add(task_row)
-
-        # When created its in pending status
-        assert task_row.status == TaskStatus.PENDING
-        assert task_row.finished_at is None
-
-        # When the status is updated to finished, the finished_at timestamp should be set
-        task_row.status = TaskStatus.FINISHED
-        database_session.add(task_row)
-        database_session.flush()
-        assert task_row.finished_at, "Should be auto generated when the status is updated to finished"
-
-        # When task status is marked as error, the finished_at timestamp should be set
-        task_row = Task(org_id=TEST_ORG_ID, task_id=random_task_id(), benchmark=benchmark_row.id)
-        task_row.status = TaskStatus.ERROR
-        database_session.add(task_row)
-        database_session.flush()
-        assert task_row.finished_at, "Should be auto generated when the status is updated to error"
-
-        # Reset the benchmark row to in progress status
-        benchmark_row.status = BenchmarkStatus.IN_PROGRESS
-        benchmark_row.finished_at = None
-        database_session.add(benchmark_row)
-        database_session.flush()
-
-        # When the benchmark status is marked as error, the finished_at timestamp should be set
-        benchmark_row.status = BenchmarkStatus.ERROR
-        database_session.add(benchmark_row)
-        database_session.flush()
-        assert benchmark_row.finished_at, "Should be auto generated when the status is updated to error"
-
-    async def test_database_relations(self, database_session: Session, example_benchmark_object: Benchmark):
-        """Verify benchmark, task, and evaluation result relationships persist correctly.
-
-        Test cases:
-        - A benchmark can be persisted and used to fetch ordered task rows.
-        - Evaluation results attach to finished tasks and update task status/timestamps.
-        """
-
-        # Add a new benchmark row to the database
-        benchmark_row = example_benchmark_object
-        database_session.add(benchmark_row)
-
-        # Can fetch it using the same id that it was created with
-        fetched_benchmark_row = database_session.get(Benchmark, benchmark_row.id)
-
-        # Base test cases
-        assert fetched_benchmark_row
-        assert fetched_benchmark_row.name == benchmark_row.name
-        assert fetched_benchmark_row.started_at is not None
-        assert fetched_benchmark_row.finished_at is None
-        assert fetched_benchmark_row.status == BenchmarkStatus.IN_PROGRESS
-
-        task_ids = [random_task_id() for _ in range(3)]
-        for task_id in task_ids:
-            _ = await self._create_task(database_session, task_id, benchmark_row.id)
-
-        # Can fetch the tasks based off the benchmark id and that the tasks are created as expected
-        fetch_tasks_query = select(Task).where(Task.benchmark == benchmark_row.id).order_by(col(Task.started_at).asc())
-        fetched_tasks = database_session.exec(fetch_tasks_query).all()
-
-        # Base test cases
-        assert len(fetched_tasks) == len(task_ids)
-        for task_id, task_row in zip(task_ids, fetched_tasks):
-            assert task_row.task_id == task_id
-            assert task_row.benchmark == benchmark_row.id
-            assert task_row.status == TaskStatus.PENDING
-            assert task_row.started_at is not None
-            assert task_row.finished_at is None
-
-        # Can add an evaluation result to the task when it is finished
-        simulated_evaluation_results: dict[str, dict[str, Any]] = {
-            task_id: {
-                "task_id": task_id,
-                "instance_id": str(uuid.uuid4()),
-                "patch_successfully_applied": True,
-                "resolved": True,
-                "resolution_status": "FULL",
-            }
-            for task_id in task_ids
-        }
-
-        for task_id, evaluation_result in simulated_evaluation_results.items():
-            # Fetch the task row based off of the human readable task_id and the benchmark foreign key
-            task_row_query = select(Task).where(Task.task_id == task_id and Task.benchmark == benchmark_row.id).limit(1)
-            task_row = database_session.exec(task_row_query).first()
-
-            assert task_row is not None
-
-            # Create the evaluation result row
-            _ = await self._create_evaluation_result(database_session, task_row, evaluation_result)
-
-        # Once the evaluation is completed, we can check if the task rows have been updated with the finished_at timestamp and status
-        fetch_tasks_query = select(Task).where(Task.benchmark == benchmark_row.id)
-        fetched_tasks = database_session.exec(fetch_tasks_query).all()
-
-        assert len(fetched_tasks) == len(task_ids)
-        for task_row in fetched_tasks:
-            assert task_row.status == TaskStatus.FINISHED
-            assert task_row.finished_at, (
-                "Finished at timestamp should be auto generated when the task status has been updated"
-            )
-
-    async def _evaluate_instance(
-        self,
-        database_session: Session,
-        benchmark_service: BenchmarkServiceClient,
-        sandbox_provider: SandboxProvider,
-        task_row: Task,
-        task_data: RetrieveTaskResponse,
-        creation_semaphore: Semaphore,
-    ) -> dict[str, str]:
-        # Change the status of the task to evaluating before we start evaluation
-        task_row.status = TaskStatus.EVALUATING
-        database_session.add(task_row)
-        database_session.flush()
-
-        async with create_sandbox(
-            sandbox_provider, task_row.task_id, task_data.source, task_data.resources, creation_semaphore
-        ) as sandbox:
-            response = await benchmark_service.setup_task(task_id=task_row.task_id, instance_id=str(sandbox.id))
-            assert response.status == "ok"
-
-            response = await benchmark_service.evaluate_instance(task_id=task_row.task_id, instance_id=sandbox.id)
-
-            return response
-
-    @pytest.mark.slow
-    async def test_end_to_end(
-        self,
-        database_session: Session,
-        benchmark_service: BenchmarkServiceClient,
-        sandbox_provider: SandboxProvider,
-        example_benchmark_object: Benchmark,
-        creation_semaphore: Semaphore,
-    ):
-        """Verify the database flow works with a live benchmark service and sandboxes.
-
-        Test cases:
-        - A sampled set of service task IDs is retrieved, evaluated, and stored in task/evaluation rows.
-        - Final score metadata and fetched evaluation results match the rows stored in the database.
-        """
-
-        try:
-            # Ensure that the benchmark service is running
-            response = await benchmark_service.health_check()
-            assert response.status == "ok"
-
-            # Create benchmark row to initiate a benchmark
-            benchmark_row = example_benchmark_object
-            database_session.add(benchmark_row)
-            database_session.flush()
-
-            # Request all of the task ids from the benchmark service
-            verify_response = await benchmark_service.verify_task_ids(task_ids=None, slice_str=None)
-
-            # Returned all of the task ids from the swebench service
-            assert verify_response.task_ids is not None
-            assert len(verify_response.task_ids) == 500
-
-            # NOTE: For testing we only are going to random sample the dataset
-            # Also to mimic limits we are going to apply a concurrency limit
-            task_ids = sample(verify_response.task_ids, 10)
-
-            logger.info(f"Sample of task ids returned: {task_ids[:10]}")
-
-            # Create the task rows for each task we are going to run
-            task_row_mapping: dict[str, Task] = {}
-            for task_id in task_ids:
-                task_row = await self._create_task(database_session, task_id, benchmark_row.id)
-                task_row_mapping[task_id] = task_row
-
-            logger.info(f"Sample of task row mapping: {str(list(task_row_mapping.values())[:250])}")
-
-            semaphore = Semaphore(5)
-            evaluation_results: dict[str, dict[str, Any] | None] = {}
-
-            async def process_task(task_id: str) -> None:
-                try:
-                    async with semaphore:
-                        task_data = await benchmark_service.retrieve_task(task_id=task_id)
-                        task_row = task_row_mapping[task_id]
-                        evaluation_result = await self._evaluate_instance(
-                            database_session,
-                            benchmark_service,
-                            sandbox_provider,
-                            task_row,
-                            task_data,
-                            creation_semaphore,
-                        )
-                        evaluation_result_row = await self._create_evaluation_result(
-                            database_session, task_row, evaluation_result
-                        )
-                        evaluation_results[task_id] = evaluation_result_row.result
-                except Exception as e:
-                    logger.error(f"Process task failed: {e}")
-
-            _ = await gather(*[process_task(task_id) for task_id in task_ids])
-
-            logger.info(f"Sample of evaluation results: {str(list(evaluation_results.values())[:100])}")
-
-            # When all tasks have been evaluated, we can make a request to get the final evaluation score
-            final_score_response = await benchmark_service.final_score(evaluation_results=evaluation_results)
-
-            logger.info(f"Final score response: {str(final_score_response)[:250]}")
-
-            # Create the final evaluation row and add it to the database
-            final_evaluation_row = await self._create_final_evaluation(
-                database_session, benchmark_row, final_score_response
-            )
-
-            # Fetch the evaluation results from the final evaluation row
-            fetched_evaluation_results = benchmark_row.fetch_evaluation_results(database_session)
-
-            results = list(fetched_evaluation_results.keys())
-            assert len(results) == len(list(evaluation_results.keys()))
-
-            logger.info(f"Sample of fetched evaluation results: {str(results[:100])}")
-
-            # check that the evaluation results retrieved match the ones stored in the database
-            for task_id, evaluation_result in fetched_evaluation_results.items():
-                evaluation_result_row = database_session.exec(
-                    select(EvaluationResult)
-                    .join(Task, col(EvaluationResult.task) == col(Task.id))
-                    .where(col(Task.task_id) == task_id)
-                ).first()
-                assert evaluation_result_row is not None
-                # fetch_evaluation_results appends agent_caused_exit_reason from the model column
-                expected = {
-                    **evaluation_result_row.result,
-                    "agent_caused_exit_reason": evaluation_result_row.agent_caused_exit_reason,
-                    "attempts": 1,
-                }
-                assert evaluation_result == expected
-
-            # Verify that the final evaluation row matches what we have in the database
-            assert final_evaluation_row.final_score == final_score_response.final_score
-            assert final_evaluation_row.properties == final_score_response.metadata
-
-        except Exception as e:
-            pytest.fail(f"End to end test failed: {e}")
